pdf_analysis/pyproject.toml
[project]
name = "pdf_anlysis"
version = "0.1.0"
description = "PDF layout analysis microservice using DocLayout-YOLO (ONNX)"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.0",
    "fastapi>=0.115.14",
    "uvicorn>=0.35.0",
    "python-multipart>=0.0.20",
    "onnxruntime>=1.18.0",
    "opencv-python-headless>=4.9.0",
    "pillow>=10.0.0",
    "PyMuPDF>=1.25.0",
    "pymupdf4llm>=1.27.2.2",
    "langchain-text-splitters>=1.1.1",
    "llama-index-core>=0.12.0",
    "transformers>=4.35.0,<5.0.0",
    # Linux は PyPI だと CUDA 付き wheel が選ばれやすく、CPU の Docker で libcusparse 等が必要になる。全環境で CPU ビルドを使う
    "torch>=2.11.0",
    "torchvision>=0.26.0",
    "einops>=0.8.2",
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.uv.sources]
torch = [{ index = "pytorch-cpu" }]
torchvision = [{ index = "pytorch-cpu" }]

[dependency-groups]
dev = [
    "mypy>=1.17.0",
    "ruff>=0.14.8",
]

[tool.mypy]
python_version = "3.13"

[[tool.mypy.overrides]]
module = [
    "fitz",
    "fitz.*",
    "onnxruntime",
    "onnxruntime.*",
    "pymupdf4llm",
    "llama_index",
    "llama_index.*",
    "transformers",
    "transformers.*",
]
ignore_missing_imports = true
